requirements.txt
pandas
seaborn
matplotlib
scikit-learn
nltk
numpy
wordcloud
gensim
sentence-transformers
hdbscan
umap-learn
tqdm
scipy
